Raise exception from try_with_exception if block failed
try_with_exception used to remove the exception completely, with this
commit if it still ends up with a failure and there was an exception it
would raise it.

There's a case when it could lead to a false positive exception, if
something failed on the first iteration, but later just failed for other
reason it would raise an original exception.

Consider this:

try_with_exception(900) { element.find('#awesome').text == 'nope' }

If it raises an ElementNotFound exception on the first iteration, but on
the second the text is not "nope": the correct response is to fail
without an exception,  but with this commit it would raise
ElementNotFound exception.